+++ QuickFIXn/ThreadedSocketReactor.cs
@@ -36,7 +36,7 @@ namespace QuickFix
         private SocketSettings socketSettings_;
         private QuickFix.Dictionary sessionDict_;
         private IPEndPoint serverSocketEndPoint_;
-        private readonly IAssumedSessionSet assumedSessionSet_;
+        private readonly AcceptorSocketDescriptor acceptorDescriptor_;
 
         #endregion
 
@@ -50,13 +50,13 @@ namespace QuickFix
         {
             
         }
-        internal ThreadedSocketReactor(IPEndPoint serverSocketEndPoint, SocketSettings socketSettings, QuickFix.Dictionary sessionDict, IAssumedSessionSet assumedSessionSet)
+        internal ThreadedSocketReactor(IPEndPoint serverSocketEndPoint, SocketSettings socketSettings, QuickFix.Dictionary sessionDict, AcceptorSocketDescriptor acceptorDescriptor)
         {
             socketSettings_ = socketSettings;
             serverSocketEndPoint_ = serverSocketEndPoint;
             tcpListener_ = new TcpListener(serverSocketEndPoint_);
             sessionDict_ = sessionDict;
-            assumedSessionSet_ = assumedSessionSet;
+            acceptorDescriptor_ = acceptorDescriptor;
         }
 
         public void Start()
@@ -115,7 +115,7 @@ namespace QuickFix
                     {
                         ApplySocketOptions(client, socketSettings_);
                         ClientHandlerThread t =
-                            new ClientHandlerThread(client, nextClientId_++, sessionDict_, socketSettings_, assumedSessionSet_);
+                            new ClientHandlerThread(client, nextClientId_++, sessionDict_, socketSettings_, acceptorDescriptor_);
                         t.Exited += OnClientHandlerThreadExited;
                         lock (sync_)
                         {
